+++ .eslintrc.cjs
@@ -5,6 +5,16 @@ module.exports = {
       extends: 'standard-with-typescript',
       parserOptions: {
         project: './tsconfig.json'
+      },
+      rules: {
+        '@typescript-eslint/strict-boolean-expressions': 'off'
+      }
+    },
+    {
+      files: ['test/**/*.spec.ts'],
+      extends: 'standard-with-typescript',
+      parserOptions: {
+        project: './tsconfig.spec.json'
       }
     }
   ]