+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # issuer-registry-client Changelog
 
+## 2.0.0 -
+
+### Changed
+- **BREAKING**: Refactor the client to use a more streamlined API, `client.load(config)`
+  and `client.didEntry(did)`
+
 ## 1.0.0 - 2022-11-30
 
 ### Added
-
 - Initial commit.